move `analyzeStylesheets` into `codemods/css/analyze.ts`
